refactor: migrate CheckboxGroup to styled components
