+++ src/components/Input/inputCheckbox/checkbox.js
@@ -1,34 +1,15 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { uniqueId } from '../../../libs/utils';
-import RenderIf from '../../RenderIf';
-import StyledContainer from '../styled/container';
-import HelpText from '../styled/helpText';
-import ErrorText from '../styled/errorText';
-import Checkbox from './checkbox';
+import Label from '../checkboxRadioLabel';
+import InlineBlockElement from '../../Structural/inlineBlockElement';
+import StyledCheckboxInput from './styled/checkbox';
 
-export default class InputCheckbox extends Component {
+export default class Checkbox extends Component {
     constructor(props) {
         super(props);
-        this.inlineTextLabelId = uniqueId('inline-text-label');
-        this.errorMessageId = uniqueId('error-message');
+        this.inputId = uniqueId('input-checkbox');
         this.inputRef = React.createRef();
-    }
-
-    getInlineTextLabelId() {
-        const { bottomHelpText } = this.props;
-        if (bottomHelpText) {
-            return this.inlineTextLabelId;
-        }
-        return undefined;
-    }
-
-    getErrorMessageId() {
-        const { error } = this.props;
-        if (error) {
-            return this.errorMessageId;
-        }
-        return undefined;
     }
 
     /**
@@ -57,30 +38,30 @@ export default class InputCheckbox extends Component {
 
     render() {
         const {
-            className,
-            style,
+            name,
             value,
             onChange,
-            label,
-            error,
-            disabled,
             tabIndex,
             onFocus,
             onBlur,
             onClick,
             onKeyDown,
-            bottomHelpText,
-            id,
-            name,
+            disabled,
             checked,
+            ariaLabelledBy,
+            ariaDescribedBy,
+            error,
+            label,
             hideLabel,
         } = this.props;
-        const errorMessageId = this.getErrorMessageId();
 
         return (
-            <StyledContainer id={id} className={className} style={style}>
-                <Checkbox
+            <InlineBlockElement>
+                <StyledCheckboxInput
+                    as="input"
+                    id={this.inputId}
                     name={name}
+                    type="checkbox"
                     value={value}
                     onChange={onChange}
                     tabIndex={tabIndex}
@@ -90,31 +71,27 @@ export default class InputCheckbox extends Component {
                     onKeyDown={onKeyDown}
                     disabled={disabled}
                     checked={checked}
-                    ariaLabelledBy={this.getInlineTextLabelId()}
-                    ariaDescribedBy={errorMessageId}
+                    aria-labelledby={ariaLabelledBy}
+                    aria-describedby={ariaDescribedBy}
                     ref={this.inputRef}
                     error={error}
+                />
+                <Label
                     label={label}
                     hideLabel={hideLabel}
+                    disabled={disabled}
+                    inputId={this.inputId}
+                    id={ariaLabelledBy}
                 />
-                <RenderIf isTrue={!!bottomHelpText}>
-                    <HelpText alignSelf="flex-start">{bottomHelpText}</HelpText>
-                </RenderIf>
-                <RenderIf isTrue={!!error}>
-                    <ErrorText alignSelf="flex-start" id={errorMessageId}>
-                        {error}
-                    </ErrorText>
-                </RenderIf>
-            </StyledContainer>
+            </InlineBlockElement>
         );
     }
 }
 
-InputCheckbox.propTypes = {
+Checkbox.propTypes = {
     value: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
     label: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
     name: PropTypes.string,
-    bottomHelpText: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
     error: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
     disabled: PropTypes.bool,
     onChange: PropTypes.func,
@@ -125,16 +102,14 @@ InputCheckbox.propTypes = {
     onKeyDown: PropTypes.func,
     checked: PropTypes.bool,
     hideLabel: PropTypes.bool,
-    className: PropTypes.string,
-    style: PropTypes.object,
-    id: PropTypes.string,
+    ariaLabelledBy: PropTypes.string,
+    ariaDescribedBy: PropTypes.string,
 };
 
-InputCheckbox.defaultProps = {
+Checkbox.defaultProps = {
     value: undefined,
     label: undefined,
     name: undefined,
-    bottomHelpText: null,
     error: null,
     disabled: false,
     onChange: () => {},
@@ -145,7 +120,6 @@ InputCheckbox.defaultProps = {
     onKeyDown: () => {},
     checked: undefined,
     hideLabel: false,
-    className: undefined,
-    style: undefined,
-    id: undefined,
+    ariaLabelledBy: undefined,
+    ariaDescribedBy: undefined,
 };
