+++ src/components/Structural/inlineBlockElement.js
@@ -2,6 +2,7 @@ import styled from 'styled-components';
 
 const InlineBlockElement = styled.div`
     display: inline-block;
+    margin-bottom: 2px;
 `;
 
 export default InlineBlockElement;
